refactored controller algorithm code
