+++ montecarlo/configs/params.yaml
@@ -56,7 +56,7 @@ true_anomaly : 0 # [deg]
 initial_attitude : [1,0,0,0] 
 initial_angular_rate : [0.2,0.3,0.2] # [rad/s]
 
-algorithm: "BaseNP" # "BaseSP" # "Lyapunov" # "Bcross" # # "BCROSS" or "PID" or "BaseSP"
+algorithm: "Bcross" # "Lyapunov" # "BaseSP" # "BaseNP" # # "BCROSS" or "PID" or "BaseSP"
 bcrossgain: 0.0028
 
 mtb_att_feedback_gains: [[0.7071, 0.0, 0.0, 0.0028, 0.0, 0.0],
